refactor(make.sh): user curl to download files instead wget

+++ make.sh
@@ -16,7 +16,7 @@ GLOBAL_BIN_DIR='/usr/local/bin'
 TO_INSTALL=('fzf' 'font')
 # Pre-Requisites as Map (command -> package)
 declare -A PRE_REQUISITES=(
-    ['wget']='wget'
+    ['curl']='curl'
     ['git']='git'
     ['unzip']='unzip'
     ['tar']='tar'
@@ -116,21 +116,26 @@ _install_nerd_font () {
     local font_name="${3:-${DEFAULT_FONT_NAME}}"
     local font_zip="${font_name}.zip"
     local font_exist="${font_name}.exist"
+    local font_url="https://github.com/ryanoasis/nerd-fonts/releases/download/latest/$font_zip"
+    local response_code='200'
 
     if [ $# -gt 3 ]; then __install_nerd_font__help $1 ; fi
     case $2 in
-        l | local)        font_path=$FONT_PATH_TO_SAVE_LOCAL ;;
-        g | global)        font_path=$FONT_PATH_TO_SAVE_GLOBAL ;;
-        *)              __install_nerd_font__help $1 ;; 
+        l | local)     font_path=$FONT_PATH_TO_SAVE_LOCAL ;;
+        g | global)    font_path=$FONT_PATH_TO_SAVE_GLOBAL ;;
+        *)             __install_nerd_font__help $1 ;;
     esac
     # check font exist or not then download and unzip
     if [ ! -f "${font_path}/${font_exist}" ]; then
         if [ ! -f "${font_path}/${font_zip}" ]; then
-            wget -nv -P "$font_path" https://github.com/ryanoasis/nerd-fonts/releases/download/latest/"$font_zip"
-        else
-            _info "Skipped: Download font '$font_name', $font_zip already exist."
+            response_code=$(curl -s -f -L "$font_url" -o "$font_path/$font_zip" -w "%{http_code}")
         fi
-        cd "$font_path" && unzip -n "$font_zip"  && rm -rf "$font_path" && fc-cache -fv && touch "$font_exist"
+        if [ $response_code == '200' ]; then
+            cd "$font_path" && unzip -n "$font_zip"  && rm -rf "$font_path" && fc-cache -fv && touch "$font_exist"
+        else
+            rm -rf "${font_path}/${font_zip}"
+            _error "Error while downloading font '$font_name', double check font name (case sensitive),\ncurl response error '$response_error'"
+        fi
     else
         _info "Skipped: Setup font '$font_name', Font already exist."
     fi
@@ -151,6 +156,8 @@ _install_fzf () {
     local bin_dir=''
     local ver='0.55.0'
     local name="fzf-${ver}-${SYS_ARCH}.tar.gz"
+    local url="https://github.com/junegunn/fzf/releases/download/v$ver/$name"
+    local response_code='200'
     
     if [ $# -ne 2 ]; then __install_fzf__help $1 ; fi
     case $2 in
@@ -160,13 +167,19 @@ _install_fzf () {
     esac
 
     if [ ! -f "${bin_dir}/$1" ]; then
-        wget -nv -P "$bin_dir" "https://github.com/junegunn/fzf/releases/download/v$ver/$name" && \
-        cd "$bin_dir" && tar xfz "$name" && rm -rf "$name" && \
-        [[ ! $PATH =~ $bin_dir ]] && \
-        _warn "'$bin_dir' didn't in PATH, Please remember to add it." && \
-        _info "use: export PATH=\"$bin_dir:\$PATH\""
+        response_code=$(curl -s -f -L "$url" -o "$bin_dir/$name" -w '%{http_code}')
+        if [ $? == 0 ]; then
+            cd "$bin_dir" && tar xfz "$name" && rm -rf "$name" && \
+            [[ ! $PATH =~ $bin_dir ]] && \
+            _warn "'$bin_dir' didn't in PATH, Please remember to add it." && \
+            _info "use: export PATH=\"$bin_dir:\$PATH\""
+        elif [ $response_code == '200' ]; then
+            _error "File Write Error: curl fail writing '$name' to '$bin_dir'"
+        else
+            _error "Invalid URL Error: curl (http_code: $response_code) fail with to download '$name' from '$url'"
+        fi
     else
-        echo "Skipped: Download $1 to bin folder '$bin_dir', $1 already exist on."
+        _info "Skipped: Download $1 to bin folder '$bin_dir', $1 already exist on."
     fi
 }
